+++ changelog.md
@@ -9,6 +9,9 @@ Please view [SomeB1oody/RustyML](https://github.com/SomeB1oody/RustyML) for more
 ### Added
 - Add Group Normalization layer
 
+### Changed
+- Refactor AdaGrad parameter update logic in convolutional layers by introducing a reusable macro to reduce duplication and improve maintainability
+
 ## [v0.9.1] - 2026-01-12 (UTC-7)
 ### Added
 - Add Instance Normalization layer